model/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
catboost==1.2.5
certifi==2024.8.30
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
feature-engine==1.8.0
fonttools==4.53.1
graphviz==0.20.3
h11==0.14.0
holidays==0.55
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==5.4.1
plotly==5.24.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rich==13.8.0
scikit-learn==1.5.1
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
tenacity==9.0.0
threadpoolctl==3.5.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
xgboost==2.0.3
tqdm
lime==0.2.0
